+++ oshi-core/src/main/java/oshi/hardware/NetworkIF.java
@@ -32,6 +32,7 @@ import org.slf4j.LoggerFactory;
 import oshi.SystemInfo;
 import oshi.hardware.platform.linux.LinuxNetworks;
 import oshi.hardware.platform.mac.MacNetworks;
+import oshi.hardware.platform.unix.solaris.SolarisNetworks;
 import oshi.hardware.platform.windows.WindowsNetworks;
 
 /**
@@ -261,6 +262,9 @@ public class NetworkIF implements Serializable {
             break;
         case MACOSX:
             MacNetworks.updateNetworkStats(this);
+            break;
+        case SOLARIS:
+            SolarisNetworks.updateNetworkStats(this);
             break;
         default:
             LOG.error("Unsupported platform. No update performed.");